% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_identifiers_collection.R
\name{identify_celltypes_using_marker}
\alias{identify_celltypes_using_marker}
\title{Run marker-based celltype identifier}
\usage{
identify_celltypes_using_marker(
  df_data,
  marker,
  method = NULL,
  N_pca = 15,
  Clustering_resolution = 1,
  N_features = 2000
)
}
\arguments{
\item{marker}{It can be either a string or a data frame (marker-matrix). For 'HiCAT' is must be a string representing the full path to the HiCAT marker file}

\item{method}{Method to use. One of 'HiCAT', 'SCINA', 'scSorter', 'scType', 'scCatch', 'Garnett'.}

\item{df_data_test}{input cell-by-gene matrix (data frame), column name must be in Hugo symbol}
}
\value{
marker-matrix in data frame
}
\description{
This function runs one of the marker-based celltype identifiers, specified by method
}
